Refactor useTenantStats hook and DashboardPage component: Updated the useTenantStats hook to accept timePeriod and an enabled flag, improving fetch control. Adjusted DashboardPage to pass the correct parameters, enhancing clarity and functionality. Removed the deprecated tenant stats API endpoint to streamline the codebase.
